Rename all files from .tsx to .js and convert their content to non-typescript
